Move IR data processing back to think
