+++ Tugboat_Arduino_3/Missions.h
@@ -1,14 +1,14 @@
 #ifndef Missions_h
 #define Missions_h
 
-#include "ArduinoComms3.h"
+#include "Sensors.h"
 
 class Missions
 {
   public:
     Missions();
 
-    RECIEVE_DATA_STRUCTURE data;
+    Sensors sensors;
 
     int tugboat_state = 0; // Which robot state to be in - for manual control, set to -1
     int cmd_heading = 0;  // Turn in degrees
